Issue 98 view edit delete job application

+++ jojo-app/src/Components/DeleteJobPopup/DeleteJobPopup.js
@@ -1,15 +1,18 @@
-import React, { useState } from "react";
+import React, { useState, useContext } from "react";
 import Modal from "react-bootstrap/Modal";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import "./DeleteJobPopup.less";
 import Alert from "react-bootstrap/Alert";
+import Cookies from "universal-cookie";
+import { userLogin } from "../../App";
 
 export default function DeleteJobPopup(props) {
-  let { show, setShowDelete, updatePosts, jobId } = props.info;
+  let { show, setShowDelete, updatePosts, jobId, myApplications } = props.info;
 
   const [showError, setShowError] = useState(false);
   const [showSuccess, setShowSuccess] = useState(false);
+  const [context, setContext] = useContext(userLogin);
 
   const handleClose = () => {
     setShowDelete(false);
@@ -22,35 +25,66 @@ export default function DeleteJobPopup(props) {
 
     e.preventDefault();
 
-    fetch(
-      "https://jobapplicationsapi.azurewebsites.net/api/JobPostsAPI/" + jobId,
-      {
-        method: "DELETE",
-      }
-    ).then((response) => {
-      if (response.ok) {
-        setShowSuccess(true);
-        updatePosts();
-      } else {
-        setShowError(true);
-      }
-    });
+    if (myApplications) {
+      const cookies = new Cookies();
+
+      fetch(
+        "https://jobapplicationsapi.azurewebsites.net/api/JobApplicantsAPI/removeJobs/" +
+          context.id +
+          "?JobId=" +
+          jobId,
+        {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: "Bearer " + cookies.get("Jwt"),
+          },
+        }
+      ).then((response) => {
+        if (response.ok) {
+          setShowSuccess(true);
+          updatePosts();
+        } else {
+          setShowError(true);
+        }
+      });
+    } else {
+      fetch(
+        "https://jobapplicationsapi.azurewebsites.net/api/JobPostsAPI/" + jobId,
+        {
+          method: "DELETE",
+        }
+      ).then((response) => {
+        if (response.ok) {
+          setShowSuccess(true);
+          updatePosts();
+        } else {
+          setShowError(true);
+        }
+      });
+    }
   };
 
   return (
     <Modal show={show} onHide={handleClose} className="popup">
       <Form onSubmit={handleEdit}>
         <Modal.Header closeButton>
-          <Modal.Title>Delete Job Posting</Modal.Title>
+          <Modal.Title>
+            {myApplications ? "Delete Application" : "Delete Job Posting"}
+          </Modal.Title>
         </Modal.Header>
-        <Modal.Body>Are you sure you want to delete this job post?</Modal.Body>
+        <Modal.Body>
+          Are you sure you want to delete
+          {myApplications ? "this application?" : " this job post?"}
+        </Modal.Body>
         <Modal.Footer>
           <Alert
             variant="success"
             className="alert"
             style={showSuccess ? { display: "block" } : { display: "none" }}
           >
-            Your job post has been deleted!
+            {myApplications ? "Your application " : "Your job post "}
+            has been deleted!
           </Alert>
           <Alert
             variant="danger"
@@ -64,7 +98,7 @@ export default function DeleteJobPopup(props) {
             Close
           </Button>
           <Button variant="primary" className="btn-danger" type="submit">
-            Delete Job
+            Delete {myApplications ? " Application" : "Job"}
           </Button>
         </Modal.Footer>
       </Form>
